Flying enemy AI
Finished basic flying enemy AI and altered enemy spawner to accommodate for the new enemy type. Also separated some code in to a different script, so multiple AI's can make use of the same function for finding all players in the scene and setting the player variable to the only active player in scene.